XML Editor
This tool makes our life simpler by representing XML file as a tree.
At a single click it provides operations like move entire sub-tree,
deep-copy, cut, edit attribute/element/data.
The interesting thing about this tool is it provides n-level of undo,
redo operations-this is implemented using 2 stacks which work
simultaneously.
Such operations like save and save as are also provided which gives nice
formatting to the file.
This tool can be used where the XML files are too large to edit or not
formatted properly.